+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Changes ###
 
-  * Event Tracker: Added 'The Big Dig' on the Map
-  * AutoButtons: Added new filter for Big Dig
-  * LocPanel: Added some new dungeon Teleports
+  * Fix chat install
+  * Created an animated Resting Indicator
+  * Added Skin for 'I Remember You'